chuj

+++ HypergraphsUI/Model/ColoringResult.cs
@@ -7,7 +7,7 @@ public class ColoringResult
     public Algorithm Algorithm { get; set; }
     public GeneratorType HypergraphType { get; set; }
     public string Size { get; set; }
-    public int AvgUsedColors { get; set; }
+    public double AvgUsedColors { get; set; }
     public int MinUsedColors { get; set; }
     public int MaxUsedColors { get; set; }
     public int Iterations { get; set; }
